requirements.txt
python-telegram-bot==13.15
langchain
langchain-groq
python-dotenv
